refactor: changed all the image names with underscore to hyphen
